# lib/api-zod/src/generated/types/uvResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SentraScope Environmental Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { UvResultSafeExposureTime } from './uvResultSafeExposureTime';
import type { UvResultSunInfo } from './uvResultSunInfo';

export interface UvResult {
  uv: number;
  uv_time: string;
  uv_max?: number;
  uv_max_time?: string;
  ozone?: number;
  ozone_time?: string;
  safe_exposure_time?: UvResultSafeExposureTime;
  sun_info?: UvResultSunInfo;
}