# shared/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SentraScope Environmental Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ConfigStatus {
  waqi: boolean;
  nasaFirms: boolean;
  openUv: boolean;
  indiaOgd: boolean;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type AirQualityDataIaqi = {[key: string]: {
  v?: number;
}};

export type AirQualityDataForecast = { [key: string]: unknown };

export type AirQualityDataAttributionsItem = { [key: string]: unknown };

export interface AirQualityData {
  aqi: number;
  city: string;
  station?: string;
  dominentpol?: string;
  time?: string;
  iaqi?: AirQualityDataIaqi;
  forecast?: AirQualityDataForecast;
  attributions?: AirQualityDataAttributionsItem[];
}

export interface AirQualityResponse {
  status: string;
  data: AirQualityData;
}

export interface FireHotspot {
  latitude: number;
  longitude: number;
  brightness?: number;
  frp?: number;
  confidence?: string;
  acq_date?: string;
  acq_time?: string;
  satellite?: string;
  instrument?: string;
  daynight?: string;
  type?: number;
}

export interface ThermalFireResponse {
  count: number;
  source?: string;
  hotspots: FireHotspot[];
}

export type UvResultSafeExposureTime = {[key: string]: number};

export type UvResultSunInfo = { [key: string]: unknown };

export interface UvResult {
  uv: number;
  uv_time: string;
  uv_max?: number;
  uv_max_time?: string;
  ozone?: number;
  ozone_time?: string;
  safe_exposure_time?: UvResultSafeExposureTime;
  sun_info?: UvResultSunInfo;
}

export interface UvResponse {
  result: UvResult;
}

export interface IndiaOgdRecord {
  state?: string;
  city?: string;
  station?: string;
  pollutant_id?: string;
  pollutant_min?: string;
  pollutant_max?: string;
  pollutant_avg?: string;
  pollutant_unit?: string;
  last_update?: string;
  [key: string]: unknown;
 }

export interface IndiaOgdResponse {
  status: string;
  total?: number;
  count?: number;
  records: IndiaOgdRecord[];
}

export type GetAirQualityParams = {
city?: string;
lat?: number;
lon?: number;
};

export type GetThermalFiresParams = {
source?: GetThermalFiresSource;
/**
 * @minimum 1
 * @maximum 10
 */
dayRange?: number;
lat?: number;
lon?: number;
};

export type GetThermalFiresSource = typeof GetThermalFiresSource[keyof typeof GetThermalFiresSource];


export const GetThermalFiresSource = {
  VIIRS_SNPP_NRT: 'VIIRS_SNPP_NRT',
  MODIS_NRT: 'MODIS_NRT',
  VIIRS_NOAA20_NRT: 'VIIRS_NOAA20_NRT',
} as const;

export type GetUvDataParams = {
lat: number;
lon: number;
alt?: number;
dt?: string;
};

export type GetIndiaOgdAirQualityParams = {
state?: string;
city?: string;
limit?: number;
};