# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SentraScope Environmental Intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: air-quality
    description: Air Quality Intelligence (WAQI)
  - name: thermal
    description: Thermal & Fire Intelligence (NASA FIRMS)
  - name: uv
    description: UV & Radiological Intelligence (OpenUV)
  - name: india-ogd
    description: India Government Open Data (CPCB)
  - name: config
    description: System configuration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /config/status:
    get:
      operationId: getConfigStatus
      tags: [config]
      summary: Get API configuration status
      description: Returns which API keys are configured
      responses:
        "200":
          description: Config status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigStatus"

  /air-quality:
    get:
      operationId: getAirQuality
      tags: [air-quality]
      summary: Get air quality data
      description: Returns air quality data for a city or lat/long
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lon
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Air quality data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirQualityResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /thermal/fires:
    get:
      operationId: getThermalFires
      tags: [thermal]
      summary: Get fire/thermal hotspot data
      parameters:
        - name: source
          in: query
          schema:
            type: string
            enum: [VIIRS_SNPP_NRT, MODIS_NRT, VIIRS_NOAA20_NRT]
        - name: dayRange
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 1
        - name: lat
          in: query
          schema:
            type: number
        - name: lon
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Thermal fire data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThermalFireResponse"

  /uv:
    get:
      operationId: getUvData
      tags: [uv]
      summary: Get UV index and ozone data
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
        - name: alt
          in: query
          schema:
            type: number
        - name: dt
          in: query
          schema:
            type: string
      responses:
        "200":
          description: UV data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UvResponse"

  /india-ogd/air-quality:
    get:
      operationId: getIndiaOgdAirQuality
      tags: [india-ogd]
      summary: Get India CPCB air quality data
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: India CPCB air quality data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndiaOgdResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ConfigStatus:
      type: object
      properties:
        waqi:
          type: boolean
        nasaFirms:
          type: boolean
        openUv:
          type: boolean
        indiaOgd:
          type: boolean
      required:
        - waqi
        - nasaFirms
        - openUv
        - indiaOgd

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    AirQualityData:
      type: object
      properties:
        aqi:
          type: number
        city:
          type: string
        station:
          type: string
        dominentpol:
          type: string
        time:
          type: string
        iaqi:
          type: object
          additionalProperties:
            type: object
            properties:
              v:
                type: number
        forecast:
          type: object
          additionalProperties: true
        attributions:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - aqi
        - city

    AirQualityResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: "#/components/schemas/AirQualityData"
      required:
        - status
        - data

    FireHotspot:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        brightness:
          type: number
        frp:
          type: number
        confidence:
          type: string
        acq_date:
          type: string
        acq_time:
          type: string
        satellite:
          type: string
        instrument:
          type: string
        daynight:
          type: string
        type:
          type: integer
      required:
        - latitude
        - longitude

    ThermalFireResponse:
      type: object
      properties:
        count:
          type: integer
        source:
          type: string
        hotspots:
          type: array
          items:
            $ref: "#/components/schemas/FireHotspot"
      required:
        - count
        - hotspots

    UvResult:
      type: object
      properties:
        uv:
          type: number
        uv_time:
          type: string
        uv_max:
          type: number
        uv_max_time:
          type: string
        ozone:
          type: number
        ozone_time:
          type: string
        safe_exposure_time:
          type: object
          additionalProperties:
            type: number
        sun_info:
          type: object
          additionalProperties: true
      required:
        - uv
        - uv_time

    UvResponse:
      type: object
      properties:
        result:
          $ref: "#/components/schemas/UvResult"
      required:
        - result

    IndiaOgdRecord:
      type: object
      properties:
        state:
          type: string
        city:
          type: string
        station:
          type: string
        pollutant_id:
          type: string
        pollutant_min:
          type: string
        pollutant_max:
          type: string
        pollutant_avg:
          type: string
        pollutant_unit:
          type: string
        last_update:
          type: string
      additionalProperties: true

    IndiaOgdResponse:
      type: object
      properties:
        status:
          type: string
        total:
          type: integer
        count:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/IndiaOgdRecord"
      required:
        - status
        - records