# shared/api-zod/src/generated/types/uvResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SentraScope Environmental Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { UvResult } from './uvResult';

export interface UvResponse {
  result: UvResult;
}